chore: remove unused code
Move PrimeBumpPolicy definition to cc file.
