+++ src/redis/util.c
@@ -164,111 +164,6 @@ int stringmatchlen(const char *pattern, int patternLen,
     return 0;
 }
 
-int stringmatch(const char *pattern, const char *string, int nocase) {
-    return stringmatchlen(pattern,strlen(pattern),string,strlen(string),nocase);
-}
-
-/* Fuzz stringmatchlen() trying to crash it with bad input. */
-int stringmatchlen_fuzz_test(void) {
-    char str[32];
-    char pat[32];
-    int cycles = 10000000;
-    int total_matches = 0;
-    while(cycles--) {
-        int strlen = rand() % sizeof(str);
-        int patlen = rand() % sizeof(pat);
-        for (int j = 0; j < strlen; j++) str[j] = rand() % 128;
-        for (int j = 0; j < patlen; j++) pat[j] = rand() % 128;
-        total_matches += stringmatchlen(pat, patlen, str, strlen, 0);
-    }
-    return total_matches;
-}
-
-/* Convert a string representing an amount of memory into the number of
- * bytes, so for instance memtoll("1Gb") will return 1073741824 that is
- * (1024*1024*1024).
- *
- * On parsing error, if *err is not NULL, it's set to 1, otherwise it's
- * set to 0. On error the function return value is 0, regardless of the
- * fact 'err' is NULL or not. */
-long long memtoll(const char *p, int *err) {
-    const char *u;
-    char buf[128];
-    long mul; /* unit multiplier */
-    long long val;
-    unsigned int digits;
-
-    if (err) *err = 0;
-
-    /* Search the first non digit character. */
-    u = p;
-    if (*u == '-') u++;
-    while(*u && isdigit(*u)) u++;
-    if (*u == '\0' || !strcasecmp(u,"b")) {
-        mul = 1;
-    } else if (!strcasecmp(u,"k")) {
-        mul = 1000;
-    } else if (!strcasecmp(u,"kb")) {
-        mul = 1024;
-    } else if (!strcasecmp(u,"m")) {
-        mul = 1000*1000;
-    } else if (!strcasecmp(u,"mb")) {
-        mul = 1024*1024;
-    } else if (!strcasecmp(u,"g")) {
-        mul = 1000L*1000*1000;
-    } else if (!strcasecmp(u,"gb")) {
-        mul = 1024L*1024*1024;
-    } else {
-        if (err) *err = 1;
-        return 0;
-    }
-
-    /* Copy the digits into a buffer, we'll use strtoll() to convert
-     * the digit (without the unit) into a number. */
-    digits = u-p;
-    if (digits >= sizeof(buf)) {
-        if (err) *err = 1;
-        return 0;
-    }
-    memcpy(buf,p,digits);
-    buf[digits] = '\0';
-
-    char *endptr;
-    errno = 0;
-    val = strtoll(buf,&endptr,10);
-    if ((val == 0 && errno == EINVAL) || *endptr != '\0') {
-        if (err) *err = 1;
-        return 0;
-    }
-    return val*mul;
-}
-
-/* Search a memory buffer for any set of bytes, like strpbrk().
- * Returns pointer to first found char or NULL.
- */
-const char *mempbrk(const char *s, size_t len, const char *chars, size_t charslen) {
-    for (size_t j = 0; j < len; j++) {
-        for (size_t n = 0; n < charslen; n++)
-            if (s[j] == chars[n]) return &s[j];
-    }
-
-    return NULL;
-}
-
-/* Modify the buffer replacing all occurrences of chars from the 'from'
- * set with the corresponding char in the 'to' set. Always returns s.
- */
-char *memmapchars(char *s, size_t len, const char *from, const char *to, size_t setlen) {
-    for (size_t j = 0; j < len; j++) {
-        for (size_t i = 0; i < setlen; i++) {
-            if (s[j] == from[i]) {
-                s[j] = to[i];
-                break;
-            }
-        }
-    }
-    return s;
-}
 /* Return the number of digits of 'v' when converted to string in radix 10.
  * See ll2string() for more information. */
 uint32_t digits10(uint64_t v) {
@@ -446,42 +341,6 @@ int string2ll(const char *s, size_t slen, long long *value) {
     return 1;
 }
 
-/* Helper function to convert a string to an unsigned long long value.
- * The function attempts to use the faster string2ll() function inside
- * Redis: if it fails, strtoull() is used instead. The function returns
- * 1 if the conversion happened successfully or 0 if the number is
- * invalid or out of range. */
-int string2ull(const char *s, unsigned long long *value) {
-    long long ll;
-    if (string2ll(s,strlen(s),&ll)) {
-        if (ll < 0) return 0; /* Negative values are out of range. */
-        *value = ll;
-        return 1;
-    }
-    errno = 0;
-    char *endptr = NULL;
-    *value = strtoull(s,&endptr,10);
-    if (errno == EINVAL || errno == ERANGE || !(*s != '\0' && *endptr == '\0'))
-        return 0; /* strtoull() failed. */
-    return 1; /* Conversion done! */
-}
-
-/* Convert a string into a long. Returns 1 if the string could be parsed into a
- * (non-overflowing) long, 0 otherwise. The value will be set to the parsed
- * value when appropriate. */
-int string2l(const char *s, size_t slen, long *lval) {
-    long long llval;
-
-    if (!string2ll(s,slen,&llval))
-        return 0;
-
-    if (llval < LONG_MIN || llval > LONG_MAX)
-        return 0;
-
-    *lval = (long)llval;
-    return 1;
-}
-
 /* Convert a string into a double. Returns 1 if the string could be parsed
  * into a (non-overflowing) double, 0 otherwise. The value will be set to
  * the parsed value when appropriate.
@@ -510,66 +369,4 @@ int string2ld(const char *s, size_t slen, long double *dp) {
 
     if (dp) *dp = value;
     return 1;
-}
-
-/* Create a string object from a long double.
- * If mode is humanfriendly it does not use exponential format and trims trailing
- * zeroes at the end (may result in loss of precision).
- * If mode is default exp format is used and the output of snprintf()
- * is not modified (may result in loss of precision).
- * If mode is hex hexadecimal format is used (no loss of precision)
- *
- * The function returns the length of the string or zero if there was not
- * enough buffer room to store it. */
-int ld2string(char *buf, size_t len, long double value, ld2string_mode mode) {
-    size_t l = 0;
-
-    if (isinf(value)) {
-        /* Libc in odd systems (Hi Solaris!) will format infinite in a
-         * different way, so better to handle it in an explicit way. */
-        if (len < 5) return 0; /* No room. 5 is "-inf\0" */
-        if (value > 0) {
-            memcpy(buf,"inf",3);
-            l = 3;
-        } else {
-            memcpy(buf,"-inf",4);
-            l = 4;
-        }
-    } else {
-        switch (mode) {
-        case LD_STR_AUTO:
-            l = snprintf(buf,len,"%.17Lg",value);
-            if (l+1 > len) return 0; /* No room. */
-            break;
-        case LD_STR_HEX:
-            l = snprintf(buf,len,"%La",value);
-            if (l+1 > len) return 0; /* No room. */
-            break;
-        case LD_STR_HUMAN:
-            /* We use 17 digits precision since with 128 bit floats that precision
-             * after rounding is able to represent most small decimal numbers in a
-             * way that is "non surprising" for the user (that is, most small
-             * decimal numbers will be represented in a way that when converted
-             * back into a string are exactly the same as what the user typed.) */
-            l = snprintf(buf,len,"%.17Lf",value);
-            if (l+1 > len) return 0; /* No room. */
-            /* Now remove trailing zeroes after the '.' */
-            if (strchr(buf,'.') != NULL) {
-                char *p = buf+l-1;
-                while(*p == '0') {
-                    p--;
-                    l--;
-                }
-                if (*p == '.') l--;
-            }
-            if (l == 2 && buf[0] == '-' && buf[1] == '0') {
-                buf[0] = '0';
-                l = 1;
-            }
-            break;
-        default: return 0; /* Invalid mode. */
-        }
-    }
-    buf[l] = '\0';
-    return l;
 }
