+++ src/redis/util.h
@@ -44,28 +44,12 @@
 #define C_OK                    0
 #define C_ERR                   -1
 
-/* long double to string conversion options */
-typedef enum {
-    LD_STR_AUTO,     /* %.17Lg */
-    LD_STR_HUMAN,    /* %.17Lf + Trimming of trailing zeros */
-    LD_STR_HEX       /* %La */
-} ld2string_mode;
-
 int stringmatchlen(const char *p, int plen, const char *s, int slen, int nocase);
-int stringmatch(const char *p, const char *s, int nocase);
-int stringmatchlen_fuzz_test(void);
-long long memtoll(const char *p, int *err);
-const char *mempbrk(const char *s, size_t len, const char *chars, size_t charslen);
-char *memmapchars(char *s, size_t len, const char *from, const char *to, size_t setlen);
-uint32_t digits10(uint64_t v);
-uint32_t sdigits10(int64_t v);
+
+
 int ll2string(char *s, size_t len, long long value);
 int string2ll(const char *s, size_t slen, long long *value);
-int string2ull(const char *s, unsigned long long *value);
-int string2l(const char *s, size_t slen, long *value);
 int string2ld(const char *s, size_t slen, long double *dp);
-int d2string(char *buf, size_t len, double value);
-int ld2string(char *buf, size_t len, long double value, ld2string_mode mode);
 
 #define LOG_MAX_LEN    1024 /* Default maximum length of syslog messages.*/
 
@@ -76,10 +60,6 @@ int ld2string(char *buf, size_t len, long double value, ld2string_mode mode);
 #define LL_WARNING 3
 #define LL_RAW (1<<10) /* Modifier to log without timestamp */
 
-
-#define LRU_BITS 24
-#define LRU_CLOCK_MAX ((1<<LRU_BITS)-1) /* Max value of obj->lru */
-#define LRU_CLOCK_RESOLUTION 1000 /* LRU clock resolution in ms */
 
 /* Bytes needed for long -> str + '\0' */
 #define LONG_STR_SIZE 21
